requirements.txt
# Flask and Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Environment and Configuration
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# Data Processing
pandas==2.1.1
numpy==1.24.3

# PDF Processing
PyPDF2==3.0.1
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.0.1

# Image Processing
opencv-python==4.8.1.78

# Database (Optional)
SQLAlchemy==2.0.21

# Data Validation
python-dateutil==2.8.2

# Utilities
certifi==2023.7.22
urllib3==2.0.4

# Development
pytest==7.4.2
black==23.7.0
flake8==6.0.0

# Production Server (Windows compatible)
waitress==2.1.2

# Google Cloud (Optional - for future enhancements)
# google-cloud-storage==2.10.0
# google-cloud-vision==3.4.4

# AI/ML (Optional - for future enhancements)
# openai==0.28.1
# transformers==4.33.2
